WIP: getting Apple Silicon enabled
