+++ .github/workflows/cmake.yml
@@ -2,7 +2,7 @@ name: CMake
 
 on:
   push:
-    branches: [ v3.68, main ]
+    branches: [ v3.69, main ]
   pull_request:
     branches: [ main ]
