+++ numeric/utils/CMakeLists.txt
@@ -1,3 +1,3 @@
 file (GLOB SOURCES "./*.cpp")
 
-compile_all("true" "utils" "Utilities" "${SOURCES}")
+compile_all("true" "utils" "Numerical Challenges/Utilities" "${SOURCES}")
